Reload when a search filter changes
Before:
When an input defined in `ajax_search_fields` changes, the select is cleared.

After:
When an input defined in `ajax_search_fields` changes, the select is cleared AND then
a new ajax request is made. If the request contain a result, the select
is populated with the result and the result is selected.